pyproject.toml
[project]
name = "urbanaut"
version = "0.0.1"
description = "Urbanaut-Club — a social platform for urban explorers, diggers, and extreme tourism enthusiasts."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.6",
    "django-cities-light>=3.10.2",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.1",
    "django-leaflet>=0.32.0",
    "django-modeltranslation>=0.19.17",
    "django-split-settings>=1.3.2",
    "django-unfold>=0.67.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-gis>=1.2.0",
    "djangorestframework-simplejwt>=5.5.1",
    "djoser>=2.3.3",
    "drf-extra-fields>=3.7.0",
    "drf-spectacular>=0.28.0",
    "drf-standardized-errors>=0.15.0",
    "geopy>=2.4.1",
    "google-genai>=1.50.0",
    "psycopg2-binary>=2.9.11",
    "python-decouple>=3.8",
    "python-dotenv>=1.2.1",
    "redis>=7.0.1",
    "requests>=2.32.5",
    "uvicorn>=0.37.0",
]

[dependency-groups]
dev = [
    "black>=25.11.0",
]
